refactor: move Game object to entity package
