+++ src/test/java/ch/uzh/ifi/hase/soprafs25/session/GameTest.java
@@ -1,5 +1,6 @@
 package ch.uzh.ifi.hase.soprafs25.session;
 
+import ch.uzh.ifi.hase.soprafs25.entity.Game;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 import java.util.Map;
